requirements.txt
selenium
pytest-html
allure-pytest